Move to `ServiceDiscovery` folder
